// src/types/adyenNexoTypes.ts
export type AdyenTerminalModel = 
  | 'S1F2' 
  | 'AMS1' 
  | 'SATURN_1000F2' 
  | 'NYC1' 
  | 'TAP_TO_PAY';